Fix double verification flow
Until now, if both types of verifications were enabled at the same time,
the user would still be redirected to the online verification page for
every action, so she wouldn't have the opportunity to see the presential
verification informational page.

Now the preferred page is the map page, and from there we'll add a link
to the online verification process.